temp update

+++ helpers/certs/generate-certs.sh
@@ -2,8 +2,10 @@
 
 cd "$(dirname "$0")"
 
-# cleanup previous artifacts
-rm -r out/ > /dev/null
+# cleanup previous client certs
+rm -r out/client > /dev/null
+
+find ./out -maxdepth 1 -type f -delete
 
 k3sFolder=$1
 
@@ -20,44 +22,56 @@ echo -n "" > ${OUTPUT_FOLDER}/index.txt
 echo -n "01" > ${OUTPUT_FOLDER}/serial
 echo -n "1000" > ${OUTPUT_FOLDER}/crlnumber
 
-mkdir -p ${OUTPUT_FOLDER}/ca/private
-mkdir -p ${OUTPUT_FOLDER}/ca/csr
-mkdir -p ${OUTPUT_FOLDER}/ca/certs
+# Only create ca if it does not exist. It allows to update the client certificates whitout having to update the ca everywhere
+if [ ! -f ${OUTPUT_FOLDER}/ca/certs/cacert.pem ]; then
 
-# generate key
-openssl genrsa -out ${OUTPUT_FOLDER}/ca/private/cakey.pem 4096
-# create CA Request
-openssl req -new -x509 -set_serial 01 -days 3650 \
-  -config ./config/openssl.cnf \
-  -extensions v3_ca \
-  -key ${OUTPUT_FOLDER}/ca/private/cakey.pem \
-  -out ${OUTPUT_FOLDER}/ca/csr/cacert.pem \
-  -subj "/C=DE/ST=Saxony/L=Dresden/O=FICODES CA/CN=FICODES-CA/serialNumber=01"
+  mkdir -p ${OUTPUT_FOLDER}/ca/private
+  mkdir -p ${OUTPUT_FOLDER}/ca/csr
+  mkdir -p ${OUTPUT_FOLDER}/ca/certs
 
-## Convert x509 CA cert
-openssl x509 -in ${OUTPUT_FOLDER}/ca/csr/cacert.pem -out ${OUTPUT_FOLDER}/ca/certs/cacert.pem -outform PEM
+  # generate key
+  openssl genrsa -out ${OUTPUT_FOLDER}/ca/private/cakey.pem 4096
+  # create CA Request
+  openssl req -new -x509 -set_serial 01 -days 3650 \
+    -config ./config/openssl.cnf \
+    -extensions v3_ca \
+    -key ${OUTPUT_FOLDER}/ca/private/cakey.pem \
+    -out ${OUTPUT_FOLDER}/ca/csr/cacert.pem \
+    -subj "/C=DE/ST=Saxony/L=Dresden/O=FICODES CA/CN=FICODES-CA/serialNumber=01"
 
-openssl pkcs8 -topk8 -nocrypt -in ${OUTPUT_FOLDER}/ca/private/cakey.pem -out ${OUTPUT_FOLDER}/ca/private/cakey-pkcs8.pem
+  ## Convert x509 CA cert
+  openssl x509 -in ${OUTPUT_FOLDER}/ca/csr/cacert.pem -out ${OUTPUT_FOLDER}/ca/certs/cacert.pem -outform PEM
 
-# Intermediate
-mkdir -p ${OUTPUT_FOLDER}/intermediate/private
-mkdir -p ${OUTPUT_FOLDER}/intermediate/csr
-mkdir -p ${OUTPUT_FOLDER}/intermediate/certs
+  openssl pkcs8 -topk8 -nocrypt -in ${OUTPUT_FOLDER}/ca/private/cakey.pem -out ${OUTPUT_FOLDER}/ca/private/cakey-pkcs8.pem
+
+else
+  echo "CA already exists, skipping generation."
+fi
+
+if [ ! -f "${OUTPUT_FOLDER}/intermediate/private/intermediate.cakey.pem" ] || [ ! -f "${OUTPUT_FOLDER}/intermediate/certs/intermediate.cacert.pem" ]; then
+
+  # Intermediate
+  mkdir -p ${OUTPUT_FOLDER}/intermediate/private
+  mkdir -p ${OUTPUT_FOLDER}/intermediate/csr
+  mkdir -p ${OUTPUT_FOLDER}/intermediate/certs
 
 
-openssl genrsa -out ${OUTPUT_FOLDER}/intermediate/private/intermediate.cakey.pem 4096
-openssl req -new -sha256 -set_serial 02 -config ./config/openssl-intermediate.cnf \
-  -subj "/C=DE/ST=Saxony/L=Dresden/O=FICODES CA/CN=FICODES-INTERMEDIATE/emailAddress=ca@ficodes.com/serialNumber=02" \
-  -key ${OUTPUT_FOLDER}/intermediate/private/intermediate.cakey.pem \
-  -out ${OUTPUT_FOLDER}/intermediate/csr/intermediate.csr.pem
-openssl ca -config ./config/openssl.cnf -extensions v3_intermediate_ca -days 2650 -notext \
-  -batch -in ${OUTPUT_FOLDER}/intermediate/csr/intermediate.csr.pem \
-  -out ${OUTPUT_FOLDER}/intermediate/certs/intermediate.cacert.pem
-openssl x509 -in ${OUTPUT_FOLDER}/intermediate/certs/intermediate.cacert.pem -out ${OUTPUT_FOLDER}/intermediate/certs/intermediate.cacert.pem -outform PEM
-openssl x509 -noout -text -in ${OUTPUT_FOLDER}/intermediate/certs/intermediate.cacert.pem
+  openssl genrsa -out ${OUTPUT_FOLDER}/intermediate/private/intermediate.cakey.pem 4096
+  openssl req -new -sha256 -set_serial 02 -config ./config/openssl-intermediate.cnf \
+    -subj "/C=DE/ST=Saxony/L=Dresden/O=FICODES CA/CN=FICODES-INTERMEDIATE/emailAddress=ca@ficodes.com/serialNumber=02" \
+    -key ${OUTPUT_FOLDER}/intermediate/private/intermediate.cakey.pem \
+    -out ${OUTPUT_FOLDER}/intermediate/csr/intermediate.csr.pem
+  openssl ca -config ./config/openssl.cnf -extensions v3_intermediate_ca -days 2650 -notext \
+    -batch -in ${OUTPUT_FOLDER}/intermediate/csr/intermediate.csr.pem \
+    -out ${OUTPUT_FOLDER}/intermediate/certs/intermediate.cacert.pem
+  openssl x509 -in ${OUTPUT_FOLDER}/intermediate/certs/intermediate.cacert.pem -out ${OUTPUT_FOLDER}/intermediate/certs/intermediate.cacert.pem -outform PEM
+  openssl x509 -noout -text -in ${OUTPUT_FOLDER}/intermediate/certs/intermediate.cacert.pem
 
-cat ${OUTPUT_FOLDER}/intermediate/certs/intermediate.cacert.pem ${OUTPUT_FOLDER}/ca/certs/cacert.pem > ${OUTPUT_FOLDER}/intermediate/certs/ca-chain-bundle.cert.pem
+  cat ${OUTPUT_FOLDER}/intermediate/certs/intermediate.cacert.pem ${OUTPUT_FOLDER}/ca/certs/cacert.pem > ${OUTPUT_FOLDER}/intermediate/certs/ca-chain-bundle.cert.pem
 
+else
+  echo "Intermediate CA already exists, skipping generation."
+fi
 
 # client
 mkdir -p ${OUTPUT_FOLDER}/client/private
